extracted theme files to a new folder
